모델 수정

+++ training/train.py
@@ -17,11 +17,9 @@ class VideoTransform:
         self.normalize = normalize
 
     def __call__(self, video):
-        # 비디오 텐서의 각 프레임에 대해 변환 적용
         transformed_frames = []
         for frame in video:
-            frame = frame.float()  # float 타입으로 변환
-            frame = self.resize(frame)
+            frame = self.resize(frame).float()  # float 타입으로 변환
             frame = self.normalize(frame)
             transformed_frames.append(frame)
         return torch.stack(transformed_frames)
@@ -31,7 +29,6 @@ transform = VideoTransform(
     normalize=Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
 )
 
-# ArgumentParser 설정
 parser = ArgumentParser()
 parser.add_argument('--batch_size', type=int, default=128)
 parser.add_argument('--lr', type=float, default=0.001)
@@ -39,19 +36,16 @@ parser.add_argument('--epochs', type=int, default=10)
 
 args = parser.parse_args()
 
-# 모델 초기화
 model = MyModel(num_classes=101).cuda()
 criterion = torch.nn.CrossEntropyLoss()
 optimizer = optim.Adam(model.parameters(), lr=args.lr)
 
-# 데이터셋과 데이터 로더 설정
 train_dataset = UCF101(root='./data/UCF-101', annotation_path='./data/annotations/ucfTrainTestlist', frames_per_clip=16, step_between_clips=1, fold=1, train=True, transform=transform)
 test_dataset = UCF101(root='./data/UCF-101', annotation_path='./data/annotations/ucfTrainTestlist', frames_per_clip=16, step_between_clips=1, fold=1, train=False, transform=transform)
 
 train_loader = DataLoader(train_dataset, batch_size=args.batch_size, shuffle=True, num_workers=4)
 test_loader = DataLoader(test_dataset, batch_size=args.batch_size, shuffle=False, num_workers=4)
 
-# 훈련 루프
 for epoch in range(args.epochs):
     model.train()
     running_loss = 0.0
@@ -65,7 +59,6 @@ for epoch in range(args.epochs):
         running_loss += loss.item()
     print(f"Epoch {epoch+1}, Loss: {running_loss/len(train_loader)}")
 
-# 평가 루프
 model.eval()
 correct = 0
 total = 0
@@ -78,4 +71,3 @@ with torch.no_grad():
         correct += (predicted == labels).sum().item()
 
 print(f"Accuracy: {100 * correct / total} %")
-
